+++ src/JudyCommon/JudyIns.c
@@ -39,7 +39,7 @@
 // population errors; see fix in 4.84:
 
 DBGCODE(extern void JudyCheckPop(Pvoid_t PArray);)
-DBGCODE(extern void JudyCheckSorted(Pjll_t Pjll, Word_t Pop1, long IndexSize);)
+DBGCODE(extern void JudyCheckSorted(Pjll_t Pjll, Word_t Pop1, intptr_t IndexSize);)
 
 #ifdef TRACEJP
 #include "JudyPrintJP.c"
